Move `get_all_tess_time_series_observations` to TessDataInterface
