+++ ramjet/photometric_database/tess_data_interface.py
@@ -1,7 +1,9 @@
 """
 Code for a class for common interfacing with TESS data, such as downloading, sorting, and manipulating.
 """
+import pandas as pd
 from astroquery.mast import Observations
+from astroquery.exceptions import TimeoutError as AstroQueryTimeoutError
 
 
 class TessDataInterface:
@@ -11,3 +13,19 @@ class TessDataInterface:
     def __init__(self):
         Observations.TIMEOUT = 1200
         Observations.PAGESIZE = 10000
+
+    @staticmethod
+    def get_all_tess_time_series_observations() -> pd.DataFrame:
+        """
+        Gets all TESS time-series observations, limited to science data product level. Repeats download attempt on
+        error.
+        """
+        tess_observations = None
+        while tess_observations is None:
+            try:
+                # noinspection SpellCheckingInspection
+                tess_observations = Observations.query_criteria(obs_collection='TESS', dataproduct_type='timeseries',
+                                                                calib_level=3)  # Science data product level.
+            except (AstroQueryTimeoutError, ConnectionError):
+                print('Error connecting to MAST. They have occasional downtime. Trying again...')
+        return tess_observations.to_pandas()
