+++ ramjet/photometric_database/tess_transit_lightcurve_label_per_time_step_database.py
@@ -63,22 +63,6 @@ class TessTransitLightcurveLabelPerTimeStepDatabase(LightcurveLabelPerTimeStepDa
         return example_path in self.meta_data_frame['lightcurve_path'].values
 
     @staticmethod
-    def get_all_tess_time_series_observations() -> pd.DataFrame:
-        """
-        Gets all TESS time-series observations, limited to science data product level. Repeats download attempt on
-        error.
-        """
-        tess_observations = None
-        while tess_observations is None:
-            try:
-                # noinspection SpellCheckingInspection
-                tess_observations = Observations.query_criteria(obs_collection='TESS', dataproduct_type='timeseries',
-                                                                calib_level=3)  # Science data product level.
-            except (AstroQueryTimeoutError, ConnectionError):
-                print('Error connecting to MAST. They have occasional downtime. Trying again...')
-        return tess_observations.to_pandas()
-
-    @staticmethod
     def get_product_list(observations: pd.DataFrame) -> pd.DataFrame:
         """
         A wrapper for MAST's `get_product_list`, allowing the use of Pandas DataFrames instead of AstroPy Tables.
